Cleanup linter warnings.
Disabled `prefer_final_locals` because it has too many occurrences and `cascade_invocations` because it was linting the tests.

Also, my team disabled the rule because the way is implemented is not really useful for us so I might rewrite it to be aligned with usages that look like builder/fluent api.