Relax void checks: allow T Function() to be assigned to void Function(). (#1115)
* Relax void checks: allow T Function() to be assigned to void Function().

See #1107. This was likely too strict and has been relaxed.

* Don't check for function refs, check for lambdas, to still lint.

Checking for everything that is a function ref is more error prone than
checking for lambdas specifically (which are easier to detect).

Rather than doing different behavior for all code which may be a
function ref, do the same code for all occurrences. And if the checked
expression is a lambda, ensure its return type is reasonable.