Fix false negatives for collection assignments
Previously the lint would allow passing a `LinkedHashSet` instantiation
if the static type wasn't `Set`, and similarly for `LinkedHashMap`. This
allowed cases where the literal would also work when passed or assigned
to something requiring any supertype like `Object`, `dynamic`, or
`Iterable`.

Invert the condition used for the `typeCheck` callback and add checks
looking specifically for the `LinkedHash` variants of the static type.
Add tests demonstrating the new places the lint will trigger.