Fix missing_whitespace_between_adjacent_strings handling of no-whitespace strings (#2735)
Fix missing_whitespace_between_adjacent_strings handling of no-whitespace strings

This involves a sizeable refactor; I pushed most of the logic of whether a
StringLiteral ends with, starts with, or contains whitespace into extensions on
StringLiteral. I think this makes a lot of sense; they're all getters that
return information about the instance.