Ignore unnecessary cast warning.
https://dart-review.googlesource.com/c/sdk/+/207360 improved the
analyzer by making CompoundAssignmentExpression a subtype of
Expression; this means that once a version of the analyzer is
published with this improvement, and the linter can depend on that
version, it will be possible to promote from Expression to
CompoundAssignmentExpression (and the linter will be able to remove a
cast).  Until then, the cast is still required to work with older
analyzers, but will trigger an "unnecessary cast" warning when
analyzing against the newest analyzer.

So for now we ignore the "unnecessary cast" warning.  Once the linter
depends on a later version of the analyzer, we'll be able to clean
this up.