Refactor rule members and binding lookups
Instead of using nullable fields, this CL creates descriptor pairs
for a type and a member, and has a native and a `dart:html` version.
Each rule uses only one descriptor, and checks to see which one it
is in the code logic when needed.

On top of this, this CL refactors the lookup methods in `dart:html`
into a class, whose instance can be used in the lints. This allows
computing the bindings lazily.