Use NamedType.name.staticElement in favor of DartType.alias.
In some circumstances, the analyzer doesn't reliably set
DartType.alias to point to the type alias that led to a type.  (This
is the ultimate cause of #3275).  I'm currently researching whether
it's possible to improve the analyzer behavior in this regard.  In the
meantime, we can fix #3275 by changing the linter to grab the type
alias element directly from NamedType.name.staticElement.

Fixes #3275.